application/app/requirements.txt
jours-feries-france==0.7.0
keras==2.11.0
numpy==1.24.0
pandas==1.5.2
psycopg2-binary==2.9.5
pytest==7.2.0
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
tensorflow==2.11.0